w2grid: add option to *not* mark search results
Marking search results is a nice feature for textual data, but when filtering your records for numerical data it is very annoying to e.g. have all 2's marked in every other column, too.

This proposal adds the option `markSearchResults` to w2grid. It defaults to `true`, so the default behaviour does not change.