fixed bug in chrome
when you don't specify an height to the form, in chrome it will not
resize and show a thin bar (2 pixels).
This is due to the abolute positioning needing to recalculate the
height.
In chrome the jquery height function give a flaot number (-0,22…) in
this case.
I just rounded the result of the jquery height function (jquery 1.10).