a crash in a XHR error handler led to a w2ui wide investigation and fix: make sure the provided callback is always passed a suitable `data` object, which is filled with a status and when the status is not a success, also an informational error message at least.
Note the TODO: given the use of eval() rather than the JSON.* system API malicious servers (e.g. servers which have been pwned or middle-manned) can execute arbitrary code on the client. This may not be desirable.

Various routes are available to solve this:

1) use JSON.* by default and stick with that. Comments in the code indicate this is not desirable either, so on to:

2) whitelist-filter: provide a custom JSON-similar parser which thus helps us restrict acceptable input to an object structure without any embedded active code a la JSON.

3) provide a user-definable hook / callback which performs the parsing; combined with (1) this might satisfy all.

searching the raw data for semicolons or other indicators of 'active code' is a blacklist-filter approach and certainly not advisable as it can be worked around (as with any blacklist filter approach anywhere)