problem with decimal/groupSeparator in italian #647
In italian we have:
"groupSymbol" : ".",
"decimalSymbol" : ",",
the opposite of english, and, in the funciont clean of w2fields,
we have the wrong order of replace for example:
if val == '123,45' that in invarant form is equal to 123.45 the function
clean first replace decimalSeparator with '.' so I have -> '123.45' and
after replace groupSeparator with '', but in italian the groupSeparator
is '.' and I have -> '12345'.
My solution is the following:
that first eliminate the groupSeparator and after replace
decimalSeparator.