allow override of edit field type on row level
I wanted to make a property list, with a property and value column. Depending on the type of property a different type of edit field should appear (i.e. text field for propery A, combo for property B). With this commit you can use an editable property on row level, the same as on column level. Row level will override the column level setting, but column level editing must be set