w2grid.js: TreeGrid: make search/expand/collapse work together properly
This is complicated, but makes the tree grid work better by:

1) Adjusting the ‘total’ field after an expand/collapse.

2) Modifying the localSearch() function to make it match hidden
(closed) children. This makes the search function more useful, and also
makes sure that a node does not suddenly disappear when we close it.

3) Call localSearch() again after expanding/collapsing a node,
otherwise “this.last.searchIds” just becomes garbage.

4) To make everything work properly, add a noReset option to localSort,
to prevent the grid to move back to row 0 when we expand a node at a
high row number with sorting enabled.

5) Sets noReset to true in the localSort call inside the add()
function: no need to school back to row 0 when rows are added to the
end of the table.