Complement search terms for enums and list
In a grid, sometimes we have fields that have a numeric value but render a text. Those numeric values normally have an associated description and is convenient to use an enum or a list to select search options for those fields. When the users want to search a value in that column, they will enter the associated text/description. So it is convenient to send the server those fields in the search record when the search term is declared also as enum or list.