Fix issue #1553
Fixed the issue where the last column disappear after reordering columns, when column header's caption contains html.

It was caused by always assuming that the parent was the column head which will not always be if the caption has one or more html nodes, and thus it couldn't get the correct position, which made the last column disappear.