Allow datetime to receive date only info to search
Allow the datetime fields to receive date only input. In that way, the remote server will properly receive the search param and can try to apply the required logic to match records in that date independently of the time. That makes easier for the user to search for info in the datetime fields.