Fixed #2012
To avoid confusion, `content` is no longer used anywhere. It is `html` also event `onContent` is renamed to `onChange`. Please note that `html` can also be an object that has `render()` method that is called to render it. It is used so that content can be a w2ui object, such as a grid, or a form.