w2utils.lang and locales reworked
w2utils.lang and locales reworked

All locales have been updated with a script to include all language settings and phrases in the same order as the "master" settings in w2utils. Phrases that existed in the locales but are no longer part of the master settings have been automatically removed, too. The order of the items in the locales is also the same as in the master settings. For "phrases" it should be alphabetically ordered, if I did everything right.

Translations added:

    'Advanced Search'
    'Do you want to delete "${item}" search item?'
    'Load ${count} more...'
    'Maximum number of files is ${count}'
    'Maximum total size is ${count}'
    'record'
    'records'
    'Request aborted.'
    'Saved Searches'
    'Search took ${count} seconds'
    'Sorting took ${count} seconds'

Translations removed:

    'Clear Search'
    'Delete Confirmation'
    'Edit Record'
    'Field should be equal to '
    'Load'
    'More'
    'Open Search Fields'
    'Records'
    'Reset Column Size'
    'Search took'
    'Search...'
    'sec'
    'Select Search Field'
    'Sorting took'
    'Toggle Line Numbers'

Translations changed:

    'Are you sure you want to delete NN records?' -> 'Are you sure you want to delete ${count} ${records}?'
    'Field should be at least NN characters.' -> 'Field should be at least ${count} characters.'
    'letters or more...' -> '${count} letters or more...'
    'Server Response' + 'sec' -> 'Server Response ${count} seconds'