Do not change grid.total if it's -1
Collapsing / Expanding children should not modify grid.total if it's -1

Reason: if grid.total is intetionally set to -1 from the server or any other time in the code, it's probably done so on purpose, e.g. to trigger the reloading of data during grid.scroll

By changing grid.total, this purposely set value is overriden.

And it's even overwritten false. Even if we only have 1 record with 3 children, when grid.total was initially -1, then adding 3 will result in grid.total becoming 2 (bit 1 + 3 is 4)

So no matter, it's a bad idea to change grid.total if it's -1